requirements.txt
flask
requests
python-dotenv
PyMuPDF
sentence-transformers
faiss-cpu
huggingface-hub
